pyproject.toml
[project]
name = "zed-toolbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cython>=3.1.6",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pyopengl>=3.1.10",
    "pyzed",
    "requests>=2.32.5",
]

[project.scripts]
zed-toolbox = "zed_toolbox:main"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
pyzed = { path = "../../../../usr/local/zed/pyzed-5.1-cp310-cp310-linux_x86_64.whl" }
